Manually create and delete the database.
Changed from having the ";create=true" string in the PU to manually doing that before getting the persistence unit. Also manually dropping the in-memory database, as just closing the EntityManagerFactory turned out to not drop the database.

The above means we can no do "update" instead of "create-drop", removing a lot of errors from the output related to non-existing tables and the like.